remove protocol, conform to Decodable, add test.
