+++ JWTDecodeTests/JWTDecodeSpec.swift
@@ -49,7 +49,7 @@ class JWTDecodeSpec: QuickSpec {
 
             it("should return original jwt string representation") {
                 let jwtString = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJjb20uc29td2hlcmUuZmFyLmJleW9uZDphcGkiLCJpc3MiOiJhdXRoMCIsInVzZXJfcm9sZSI6ImFkbWluIn0.sS84motSLj9HNTgrCPcAjgZIQ99jXNN7_W9fEIIfxz0"
-                let jwt = try! decode(jwt: jwtString)
+                let jwt = try! JWT(jwtString)
                 expect(jwt.string).to(equal(jwtString))
             }
 
@@ -59,12 +59,12 @@ class JWTDecodeSpec: QuickSpec {
 
             it("should decode valid jwt") {
                 let jwt = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJjb20uc29td2hlcmUuZmFyLmJleW9uZDphcGkiLCJpc3MiOiJhdXRoMCIsInVzZXJfcm9sZSI6ImFkbWluIn0.sS84motSLj9HNTgrCPcAjgZIQ99jXNN7_W9fEIIfxz0"
-                expect(try! decode(jwt: jwt)).toNot(beNil())
+                expect(try! JWT(jwt)).toNot(beNil())
             }
 
             it("should raise exception with invalid json in jwt") {
                 let token = "HEADER.BODY.SIGNATURE"
-                expect { try decode(jwt: token) }
+                expect { try JWT(token) }
                     .to(throwError { (error: Error) in
                         expect(error).to(beDecodeErrorWithCode(.invalidJSON("HEADER")))
                     })
@@ -72,16 +72,28 @@ class JWTDecodeSpec: QuickSpec {
 
             it("should raise exception with missing parts") {
                 let token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJzdWIifQ"
-                expect { try decode(jwt: token) }
+                expect { try JWT(token) }
                     .to(throwError { (error: Error) in
                         expect(error).to(beDecodeErrorWithCode(.invalidPartCount(token, 2)))
                     })
             }
 
         }
+        describe("decodable") {
+            struct JWTTestDecodable: Decodable {
+                let jwt: JWT
+            }
+            
+            it("should parse a json string to jwt") {
+                let jsonData = "{\"jwt\":\"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJjb20uc29td2hlcmUuZmFyLmJleW9uZDphcGkiLCJpc3MiOiJhdXRoMCIsInVzZXJfcm9sZSI6ImFkbWluIn0.sS84motSLj9HNTgrCPcAjgZIQ99jXNN7_W9fEIIfxz0\"}".data(using: .utf8)!
+                let jsonDecoder = JSONDecoder()
+                expect { try jsonDecoder.decode(JWTTestDecodable.self, from: jsonData) }
+                    .toNot(throwError())
+            }
+        }
 
         describe("jwt parts") {
-            let jwt = try! decode(jwt: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJzdWIifQ.xXcD7WOvUDHJ94E6aVHYgXdsJHLl2oW7ZXm4QpVvXnY")
+            let jwt = try! JWT("eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJzdWIifQ.xXcD7WOvUDHJ94E6aVHYgXdsJHLl2oW7ZXm4QpVvXnY")
 
             it("should return header") {
                 expect(jwt.header as? [String: String]).to(equal(["alg": "HS256", "typ": "JWT"]))
@@ -122,7 +134,7 @@ class JWTDecodeSpec: QuickSpec {
 
             describe("registered claims") {
 
-                let jwt = try! decode(jwt: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJodHRwczovL3NhbXBsZXMuYXV0aDAuY29tIiwic3ViIjoiYXV0aDB8MTAxMDEwMTAxMCIsImF1ZCI6Imh0dHBzOi8vc2FtcGxlcy5hdXRoMC5jb20iLCJleHAiOjEzNzI2NzQzMzYsImlhdCI6MTM3MjYzODMzNiwianRpIjoicXdlcnR5MTIzNDU2IiwibmJmIjoxMzcyNjM4MzM2fQ.LvF9wSheCB5xarpydmurWgi9NOZkdES5AbNb_UWk9Ew")
+                let jwt = try! JWT("eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJodHRwczovL3NhbXBsZXMuYXV0aDAuY29tIiwic3ViIjoiYXV0aDB8MTAxMDEwMTAxMCIsImF1ZCI6Imh0dHBzOi8vc2FtcGxlcy5hdXRoMC5jb20iLCJleHAiOjEzNzI2NzQzMzYsImlhdCI6MTM3MjYzODMzNiwianRpIjoicXdlcnR5MTIzNDU2IiwibmJmIjoxMzcyNjM4MzM2fQ.LvF9wSheCB5xarpydmurWgi9NOZkdES5AbNb_UWk9Ew")
 
 
                 it("should return issuer") {
@@ -139,7 +151,7 @@ class JWTDecodeSpec: QuickSpec {
 
                 context("multiple audiences") {
 
-                    let jwt = try! decode(jwt: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhdWQiOlsiaHR0cHM6Ly9zYW1wbGVzLmF1dGgwLmNvbSIsImh0dHBzOi8vYXBpLnNhbXBsZXMuYXV0aDAuY29tIl19.cfWFPuJbQ7NToa-BjHgHD1tHn3P2tOP5wTQaZc1qg6M")
+                    let jwt = try! JWT("eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhdWQiOlsiaHR0cHM6Ly9zYW1wbGVzLmF1dGgwLmNvbSIsImh0dHBzOi8vYXBpLnNhbXBsZXMuYXV0aDAuY29tIl19.cfWFPuJbQ7NToa-BjHgHD1tHn3P2tOP5wTQaZc1qg6M")
 
                     it("should return all audiences") {
                         expect(jwt.audience).to(equal(["https://samples.auth0.com", "https://api.samples.auth0.com"]))
